move frustum culling lock types and gpu profiler wrapper out of ui module
