+++ ikari/src/ui.rs
@@ -8,115 +8,8 @@ use iced_winit::{runtime, Clipboard, Viewport};
 
 use winit::{event::WindowEvent, window::Window};
 
-use crate::renderer::CullingFrustumLock;
-
 #[derive(Debug, Clone, Default, PartialEq, Eq)]
 pub struct EmptyUiOverlay;
-
-// TODO: move out of ui module (but keep them in ikari?)
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Default)]
-pub enum CullingFrustumLockMode {
-    Full,
-    FocalPoint,
-    #[default]
-    None,
-}
-
-// TODO: move out of ui module (but keep them in ikari?)
-impl CullingFrustumLockMode {
-    pub const ALL: [CullingFrustumLockMode; 3] = [
-        CullingFrustumLockMode::None,
-        CullingFrustumLockMode::Full,
-        CullingFrustumLockMode::FocalPoint,
-    ];
-}
-
-// TODO: move out of ui module (but keep them in ikari?)
-impl From<CullingFrustumLock> for CullingFrustumLockMode {
-    fn from(value: CullingFrustumLock) -> Self {
-        match value {
-            CullingFrustumLock::Full(_) => CullingFrustumLockMode::Full,
-            CullingFrustumLock::FocalPoint(_) => CullingFrustumLockMode::FocalPoint,
-            CullingFrustumLock::None => CullingFrustumLockMode::None,
-        }
-    }
-}
-
-// TODO: move out of ui module (but keep them in ikari?)
-impl std::fmt::Display for CullingFrustumLockMode {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "{}",
-            match self {
-                CullingFrustumLockMode::Full => "On",
-                CullingFrustumLockMode::FocalPoint => "Only FocalPoint",
-                CullingFrustumLockMode::None => "Off",
-            }
-        )
-    }
-}
-
-// TODO: move out of ui module (but keep them in ikari?)
-pub struct GpuTimerScopeResultWrapper(pub wgpu_profiler::GpuTimerScopeResult);
-
-// TODO: move out of ui module (but keep them in ikari?)
-impl std::ops::Deref for GpuTimerScopeResultWrapper {
-    type Target = wgpu_profiler::GpuTimerScopeResult;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-// TODO: move out of ui module (but keep them in ikari?)
-impl std::fmt::Debug for GpuTimerScopeResultWrapper {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "(TODO implement better formatter for type) {:?} -> {:?}",
-            self.label, self.time
-        )
-    }
-}
-
-// TODO: move out of ui module (but keep them in ikari?)
-impl Clone for GpuTimerScopeResultWrapper {
-    fn clone(&self) -> Self {
-        Self(wgpu_profiler::GpuTimerScopeResult {
-            label: self.label.clone(),
-            time: self.time.clone(),
-            nested_scopes: clone_nested_scopes(&self.nested_scopes),
-            pid: self.pid,
-            tid: self.tid,
-        })
-    }
-}
-
-// TODO: move out of ui module (but keep them in ikari?)
-fn clone_nested_scopes(
-    nested_scopes: &[wgpu_profiler::GpuTimerScopeResult],
-) -> Vec<wgpu_profiler::GpuTimerScopeResult> {
-    nested_scopes
-        .iter()
-        .map(|nested_scope| wgpu_profiler::GpuTimerScopeResult {
-            label: nested_scope.label.clone(),
-            time: nested_scope.time.clone(),
-            nested_scopes: clone_nested_scopes(&nested_scope.nested_scopes),
-            pid: nested_scope.pid,
-            tid: nested_scope.tid,
-        })
-        .collect()
-}
-
-// TODO: move out of ui module (but keep them in ikari?)
-pub fn collect_frame_time_ms(frame_times: &Vec<GpuTimerScopeResultWrapper>) -> f64 {
-    let mut result = 0.0;
-    for frame_time in frame_times {
-        result += (frame_time.time.end - frame_time.time.start) * 1000.0;
-    }
-    result
-}
 
 // integrates iced into ikari
 // based off of https://github.com/iced-rs/iced/tree/master/examples/integration_wgpu
